thermal: consistently use int for temperatures
The thermal code uses int, long and unsigned long for temperatures
in different places.

Using an unsigned type limits the thermal framework to positive
temperatures without need. Also several drivers currently will report
temperatures near UINT_MAX for temperatures below 0°C. This will probably
immediately shut the machine down due to overtemperature if started below
0°C.

'long' is 64bit on several architectures. This is not needed since INT_MAX °mC
is above the melting point of all known materials.

Consistently use a plain 'int' for temperatures throughout the thermal code and
the drivers. This only changes the places in the drivers where the temperature
is passed around as pointer, when drivers internally use another type this is
not changed.

Acked-by: Geert Uytterhoeven <geert+renesas@glider.be>
Cc: Punit Agrawal <punit.agrawal@arm.com>
Cc: Zhang Rui <rui.zhang@intel.com>
Cc: Eduardo Valentin <edubezval@gmail.com>
Cc: linux-pm@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: Jean Delvare <jdelvare@suse.de>
Cc: Peter Feuerer <peter@piie.net>
Cc: Heiko Stuebner <heiko@sntech.de>
Cc: Lukasz Majewski <l.majewski@samsung.com>
Cc: Stephen Warren <swarren@wwwdotorg.org>
Cc: Thierry Reding <thierry.reding@gmail.com>
Cc: linux-acpi@vger.kernel.org
Cc: platform-driver-x86@vger.kernel.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-omap@vger.kernel.org
Cc: linux-samsung-soc@vger.kernel.org
Cc: Guenter Roeck <linux@roeck-us.net>
Cc: Rafael J. Wysocki <rjw@rjwysocki.net>
Cc: Maxime Ripard <maxime.ripard@free-electrons.com>
Cc: Darren Hart <dvhart@infradead.org>
Cc: lm-sensors@lm-sensors.org

+++ drivers/thermal/cpu_cooling.c
@@ -42,9 +42,9 @@ extern bool is_spm_mode_enabled(void);
 
 u32 profile_freq[NUM_CLUSTERS];
 int get_profile_power(enum ipa_actor actor, unsigned int *power);
-int hisi_calc_static_power(const struct cpumask *cpumask, unsigned long temp,
+int hisi_calc_static_power(const struct cpumask *cpumask, int temp,
 				unsigned long u_volt, u32 *static_power);
-int get_soc_target_temp(struct thermal_cooling_device *cdev, unsigned long *target_temp);
+int get_soc_target_temp(struct thermal_cooling_device *cdev, int *target_temp);
 #endif
 extern unsigned int g_ipa_freq_limit[];
 #endif
@@ -906,7 +906,7 @@ static int cpufreq_freq2volt(struct cpufreq_cooling_device *cpufreq_device, unsi
 
 static int cpufreq_power2freq(struct thermal_cooling_device *cdev, u32 power, u32 *freq)
 {
-	unsigned long target_temp = 0;
+	int target_temp = 0;
 	struct cpufreq_cooling_device *cpufreq_device = cdev->devdata;
 	struct power_table *pt = cpufreq_device->dyn_power_table;
 	int i, ret;
